Struct whiteread::WhiteReader

pub struct WhiteReader<B: BufRead> {
    // some fields omitted
}

Wrapper for BufRead allowing easy parsing values from a Reader.

This struct contains line-buffer, which enables scanf-like behavior (newline-agnostic parsing) and also provides almost zero-allocation parsing.

Examples

The following will accept both of the following inputs:

1

2

1 2

let i = std::io::stdin();
let mut i = WhiteReader::new(i.lock());
let (n, k): (i32, f64) = i.p();

Methods

impl<B: BufRead> WhiteReader<B>

fn new(buf: B) -> WhiteReader<B>

Wraps a BufRead.

Note that you don't have to pass an owned buffered reader, it could be also &mut`&mut`.

fn parse<T: White>(&mut self) -> WhiteResult<T>

Parses a White value without specialy treating newlines (just like scanf`scanfor` or cin>>`cin>>`)

fn p<T: White>(&mut self) -> T

Just parse().unwrap().

Use it if you really value your time. ;)

fn get_line(&mut self) -> WhiteResult<&str>

Reads a new line and returns it.

This functions should be used when parse`parse-like functions are insufficient or just to get a preview of a line. Note that line's content will not be considered consumed and will be available forparseand` and continue_line.

fn parse_line<T: White>(&mut self) -> WhiteResult<T>

Reads a new line from input and parses it into White value.

It's not an error when there are leftovers. They will be available to parse after. For more fine-grained control use get_line`get_line`.

fn continue_line<T: White>(&mut self) -> WhiteResult<T>

Parses remaining part of line into White value.

fn unwrap(self) -> B

Gets underlying buffer back.