Enum whiteread::WhiteError

pub enum WhiteError {
    TooShort,
    Leftovers,
    ParseError,
    IoError(Error),
}

Error which can occur while parsing White`White` object.

It's convertible into io::Error, so it composes well with other reading functions.

Examples

if let Err(TooShort) = parse_string::<(u8, u16)>("1") {} else { panic!(); }
if let Err(Leftovers) = parse_string::<char>("x y z") {} else { panic!(); }
if let Err(ParseError) = parse_string::<i32>("seven") {} else { panic!(); }

Variants

TooShort

There was not enough input to parse a value.

Leftovers

Excessive input was provided.

ParseError

Parse error occured (data was in invalid format).

IoError

IO Error occured.