Struct whiteread::WhiteReader

pub struct WhiteReader<B: BufRead> {
    // some fields omitted
}

Wrapper for BufRead allowing easy parsing values from a Reader.

This struct contains line-buffer, which enables scanf-like behavior (newline-agnostic parsing) and also provides almost zero-allocation parsing.

Examples

This code

let i = std::io::stdin();
let mut i = WhiteReader::new(i.lock());
let (n, k): (i32, f64) = i.p();

will accept both of following inputs:

1

2

1 2

Overview of how various methods handle newlines:

let data = std::io::Cursor::new(b"1 2\n\n3 4 5\n6 7\n8\n" as &[u8]);
let mut r = WhiteReader::new(data);
assert_eq!(r.next_line().unwrap().trim(), "1 2");
assert_eq!(r.parse().ok(), Some(1));
assert_eq!(r.parse().ok(), Some( (2, 3) ));   // continue_line would return `TooShort` here
assert_eq!(r.continue_line().ok(), Some(4)); // finish_line would return `Leftovers` here
assert_eq!(r.start_line().ok(), Some(6));   // line would return `Leftovers` here
assert_eq!(r.line().ok(), Some(8));
// from now, everything will return Err(TooShort)

Methods

impl<B: BufRead> WhiteReader<B>

Constructors

fn new(buf: B) -> WhiteReader<B>

Wraps a BufRead.

Note that you don't have to pass an owned buffered reader, it could be also &mut`&mut`.

impl<B: BufRead> WhiteReader<B>

Parsing methods

Following methods parse some part of input into a White value.

Errors

These methods may return TooShort`TooShort,`, ParseError or IoError`IoError` error variant. If they return other variants too, it is stated explicitely.

fn parse<T: White>(&mut self) -> WhiteResult<T>

Parses a White value without specialy treating newlines (just like scanf`scanfor` or cin>>`cin>>`)

fn p<T: White>(&mut self) -> T

Just parse().unwrap().

Use it if you really value your time. ;)

fn line<T: White>(&mut self) -> WhiteResult<T>

Reads a new line from input and parses it into White value as a whole.

The function is called just line`linefor brevity and also to make it look different than globalread_line` to avoid mistakes.

Errors

Additionaly to usual parse errors, this method may also return Leftovers.

fn start_line<T: White>(&mut self) -> WhiteResult<T>

Reads a new line from input and parses some part of it into White value.

fn continue_line<T: White>(&mut self) -> WhiteResult<T>

Parses some part of current line into White value.

fn finish_line<T: White>(&mut self) -> WhiteResult<T>

Parses remaining part of current line into White value.

It could be used with T=()`T=()`, to just check if we're on the end of line.

Errors

Additionaly to usual parse errors, this method may also return Leftovers.

impl<B: BufRead> WhiteReader<B>

Additional methods

fn next_line(&mut self) -> WhiteResult<&str>

Reads a new line and returns it.

This function should be used when parse`parse-like functions are insufficient or just to get a preview of a line. Note that line's content will not be considered consumed and will be available forparseand` and continue_line.

fn unwrap(self) -> B

Gets underlying buffer back.